[Refactoring] Code format in project
