+++ src/main/java/ua/strongBody/dao/BookingDAO.java
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.UUID;
 
 public interface BookingDAO extends GeneralDAO<Booking> {
+
     void saveWithoutId(Booking booking);
 
     List<Booking> getBookingsByCartId(UUID cartId);
