+++ src/main/java/ua/strongBody/dao/CartDAO.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import java.util.UUID;
 
 public interface CartDAO extends GeneralDAO<Cart> {
+
     void saveWithoutId(Cart cart);
 
     Optional<Cart> findCartByCustomerId(UUID customerId);
