+++ src/main/java/ua/strongBody/dao/CustomerDAO.java
@@ -6,6 +6,7 @@ import ua.strongBody.models.Customer;
 import java.util.Optional;
 
 public interface CustomerDAO extends GeneralDAO<Customer> {
+
     Optional<Customer> findFirstByUsername(String username);
 
     void saveWithoutId(Customer customer) throws DataAccessException;
