+++ src/main/java/ua/strongBody/dao/OrderDAO.java
@@ -6,5 +6,6 @@ import java.util.List;
 import java.util.UUID;
 
 public interface OrderDAO extends GeneralDAO<Order> {
+
     List<Order> findOrdersByCustomerId(UUID customerId);
 }
