+++ src/main/java/ua/strongBody/dao/ProductDAO.java
@@ -5,6 +5,7 @@ import ua.strongBody.models.Product;
 import java.util.Optional;
 
 public interface ProductDAO extends GeneralDAO<Product> {
+
     Optional<Product> findByArticle(String article);
 
     void saveWithoutId(Product product);
