+++ src/main/java/ua/strongBody/exceptions/FieldNotFoundException.java
@@ -1,6 +1,7 @@
 package ua.strongBody.exceptions;
 
 public class FieldNotFoundException extends RuntimeException {
+
     public FieldNotFoundException(String message) {
         super(message);
     }
