+++ src/main/java/ua/strongBody/exceptions/ValidationException.java
@@ -1,6 +1,7 @@
 package ua.strongBody.exceptions;
 
 public class ValidationException extends Exception {
+
     public ValidationException(String message) {
         super(message);
     }
