+++ src/main/java/ua/strongBody/facade/OrderFacade.java
@@ -5,6 +5,7 @@ import ua.strongBody.models.Order;
 import java.util.List;
 
 public interface OrderFacade {
+
     void createOrderFromCartByCustomerUsername(String username);
 
     List<Order> getOrdersByCustomerUsername(String username);
