+++ src/main/java/ua/strongBody/facade/impl/OrderFacadeImpl.java
@@ -1,6 +1,7 @@
-package ua.strongBody.facade;
+package ua.strongBody.facade.impl;
 
 import org.springframework.stereotype.Component;
+import ua.strongBody.facade.OrderFacade;
 import ua.strongBody.models.Cart;
 import ua.strongBody.models.Customer;
 import ua.strongBody.models.Order;
