+++ src/main/java/ua/strongBody/models/State.java
@@ -1,5 +1,5 @@
 package ua.strongBody.models;
 
 public enum State {
-    ACTIVE , BANNED , DELETED;
+    ACTIVE, BANNED, DELETED;
 }
