+++ src/main/java/ua/strongBody/models/forms/RegistrationForm.java
@@ -5,6 +5,7 @@ import ua.strongBody.models.Product;
 import java.util.StringJoiner;
 
 public class RegistrationForm {
+
     private String firstName;
     private String lastName;
     private String username;
