+++ src/main/java/ua/strongBody/populator/Populator.java
@@ -1,5 +1,6 @@
 package ua.strongBody.populator;
 
 public interface Populator<T, V> {
+
     void convert(T t, V v);
 }
