+++ src/main/java/ua/strongBody/processors/post/PostProcessor.java
@@ -3,6 +3,7 @@ package ua.strongBody.processors.post;
 import java.util.List;
 
 public interface PostProcessor<T> {
+
     void postProcess(List<T> tList);
 
     void postProcess(T t);
