+++ src/main/java/ua/strongBody/services/BookingService.java
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.UUID;
 
 public interface BookingService extends GeneralService<Booking> {
+
     void createBooking(Booking booking);
 
     List<Booking> getCustomerBookingsByCartId(UUID cartId);
