+++ src/main/java/ua/strongBody/services/CartService.java
@@ -5,5 +5,6 @@ import ua.strongBody.models.Cart;
 import java.util.UUID;
 
 public interface CartService extends GeneralService<Cart> {
+
     Cart findCartByCustomerId(UUID customerId);
 }
