+++ src/main/java/ua/strongBody/services/CustomerService.java
@@ -6,5 +6,4 @@ import ua.strongBody.models.Customer;
 public interface CustomerService extends GeneralService<Customer> {
 
     Customer findByUsername(String username) throws FieldNotFoundException;
-
 }
