+++ src/main/java/ua/strongBody/services/OrderService.java
@@ -6,5 +6,6 @@ import java.util.List;
 import java.util.UUID;
 
 public interface OrderService extends GeneralService<Order> {
+
     List<Order> findOrdersByCustomerId(UUID customerId);
 }
