+++ src/main/java/ua/strongBody/services/ProductService.java
@@ -3,5 +3,6 @@ package ua.strongBody.services;
 import ua.strongBody.models.Product;
 
 public interface ProductService extends GeneralService<Product> {
+
     void createProduct(Product product);
 }
