+++ src/main/java/ua/strongBody/services/RegistrationService.java
@@ -3,5 +3,6 @@ package ua.strongBody.services;
 import ua.strongBody.models.forms.RegistrationForm;
 
 public interface RegistrationService {
+
     boolean register(RegistrationForm registrationForm);
 }
