+++ src/main/java/ua/strongBody/validation/Validator.java
@@ -3,5 +3,6 @@ package ua.strongBody.validation;
 import ua.strongBody.exceptions.ValidationException;
 
 public interface Validator<V> {
+
     void validate(V v) throws ValidationException;
 }
