+++ src/test/java/ua/strongBody/facade/OrderFacadeImplTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import ua.strongBody.facade.impl.OrderFacadeImpl;
 import ua.strongBody.models.Cart;
 import ua.strongBody.models.Customer;
 import ua.strongBody.models.Order;
